==> step 16 ssGSEA drafting.R <==
devtools::install_github('dviraran/xCell')


library(GSVA)
library(limma)
library(GSEABase)
setwd("E:\\ÎÄÕÂ\\ÎÄÕÂ\\»µËÀ+·ÎÏÙ°©\\132pyroptosis\\32.ssGSEA\\try")          


immuneScore=function(expFile=null, gmtFile=null, project=null){

	rt=read.table(expFile, header=T, sep="\t", check.names=F)
	rt=as.matrix(rt)
	rownames(rt)=rt[,1]
	exp=rt[,2:ncol(rt)]
	dimnames=list(rownames(exp),colnames(exp))
	mat=matrix(as.numeric(as.matrix(exp)),nrow=nrow(exp),dimnames=dimnames)
	mat=avereps(mat)
	mat=mat[rowMeans(mat)>0,]
	

	geneSet=getGmt(gmtFile, geneIdType=SymbolIdentifier())
	

	ssgseaScore=gsva(mat, geneSet, method='ssgsea', kcdf='Gaussian', abs.ranking=TRUE)

	normalize=function(x){
	  return((x-min(x))/(max(x)-min(x)))}

	ssgseaOut=normalize(ssgseaScore)
	ssgseaOut=rbind(id=colnames(ssgseaOut),ssgseaOut)
	write.table(ssgseaOut, file=paste0(project, ".score.txt"), sep="\t", quote=F, col.names=F)
}


immuneScore(expFile="TCGA.normalize.txt", gmtFile="immune.gmt", project="TCGA")
immuneScore(expFile="GEO.normalize.txt", gmtFile="immune.gmt", project="GEO")
